Make static pages cacheable for non-logged in users
While testing with integration CDN I noticed that the cached static
pages remained as cached items even once a user was logged in - which
was not intentional.

To fix this I've made a couple of changes to how the caching behaves:

1) we don't cache when a user is logged in - which simplifies that side
   of the issue
2) we set a vary header of Cookie so that a change in cookies is
   reflected in any cache hashing. This is so that if a user logs in and
   then has a session cookie they won't be served a cached version of
   this page by either their browser or Fastly

Using a Vary header of Cookie is controversial [1] as it can be of such
limited value. However because we strip all cookies if there isn't a
session cookie [2] I understand that this will act as a toggle to fastly
between all logged in and all users not logged in.

[1]: https://www.fastly.com/blog/best-practices-using-vary-header/#vary-cookie
[2]: https://github.com/alphagov/govuk-fastly/blob/3b4c8ec6184318e7a1c08925ad1084ea42e88ff5/www/www.vcl.tftpl#L379-L383