Update landing page content
This commit updates the early access landing page’s label and info text. One thing to note is that I’ve moved the aria-describedby attribute from the blue start button to the input field because I believe an aria-describedby attribute on an input element is announced more reliably. For context, during DAC’s live screen reader demo, it was noted that the informational text wasn’t always announced when referenced via aria-describedby on the blue start button. I’ve created a card to look into this later but I’ve tried to fix things in this specific scenario as it’s important information that we don’t want missed.