Add margin: 0 to input
This is to fix an issue caused by removing align-items: center in PR #650 - this was originally introduced in PR #403 in order to resolve an issue seen on some iOS devices where the input and submit button were slightly misaligned. By chance, however, I found that it caused the issue in PR #650. Upon further investigation in browserstack, it looked like the input element was getting a margin-top: 2px and a margin-bottom: 2px applied to it in some devices (I've tried to demonstrate the issue with some screenshots but I'm happy to do a real time demo if it's confusing/would be easier). I'm not sure why the 2px worth of margin is being applied but adding margin: 0 to the input seems to do the trick and keeps everything aligned. I'm quite confident that this won't have any adverse effects as we weren't setting any margin on the input in the first place and I didn't see any issues from testing multiple devices and browsers in browserstack.