Add sign out link to header
This adds the link to the header so a user can sign out.

This was mostly straightfoward, except that for users with JS enabled,
we're injecting a print button into the nav list at the bottom of the
list, whereas the designs call for the sign out button to always be
last.

So to address this, we give the new sign out nav item a data attribute
of `after-print=true`. We can then use this attribute in the header JS
to find the element and insert the print button directly before this
element.

If we can't find the element, we just stick it at the end like normal.

A minor niggle here is that we're applying the data attribute to the
`a` in the header, and it'd be neater to apply it to the `li` instead.
But because we're using the `navigation_items` from publishing
components, the only available option is to add the data attribute to
the link. It just means in the JS we have to find the link,
then find the parent to get the `li` to insert the print item before.