Return remaining question count when updating conversation
We want to show a hint to the user if they're approaching their limit on
number of questions they can ask.

To do that, we'll return the number of remaining questions they can ask
in the JSON response when they ask a question.

The logic is a little more nuanced though. We only want to show a hint
to the user if their remaining questions is within some threshold. This
is configured in the app to be 20.

So if the count is within this threshold, we return the remaining
question count, otherwise we return an empty string. The FE will then be
extended to show the hint to the user if this value is present.