Show question remaining hint
When we get the JSON response from asking a question, it comes with a
`show_remaining_questions` attribute.

This attribute can either be `null` or an integer describing the number
of questions a user has left to ask.

If it's an integer, we want to inform the user by showing a hint just
above the question input field.

So we pass this value to the `QuestionForm` component in the event
detail as this component is responsible for updating the UI.

The component rendered in the `_question_form` partial is given an ID so
we can query it in the JS, and a blank value for `text` initially, as we
don't want to show it on first page load. Ideally it'd be good not to
have to set this empty string, but the component requires some value.