Add rake task to bump places settings
This rake task will be used to automate bumping the places settings
for instant access places and delayed access places.

The rake task takes a places_type argument. This must either be "instant_access"
or "delayed_access". If not we exit the rake task.

We use the places_type to retrieve the config for the rake task from
either config.increment_instant_access_places or
config.decrement_instant_access_places.

Both of these provide 4 values:

- not_before: (date)
- not_after: (date)
- places: (integer)
- max_places: (integer)

These are used to determine whether to bump the setting or not.

If places is not defined in the config or it's not a positive integer we
abort with an error and message.

If the rake task runs before the 'not_before' date or after the 'not_after' date
then the rake task will exit (no error).

If the rake task runs and the instant access places/delayed are already
at the 'max_places' in the config then we output that the places
are at the max places and exit (no error).

If the setting is below the max_places it will either

- increment the places by the places amount in the config
- increment the places to the max_places if the increment would take
  the places above the max_places

Prior to merge i'll need to confirm the values we want to use for each
of these values, but for now i've set some placeholder values.