Remove flipper gem and associated code
We’ve not got a use case for flipper at the moment and it seems
unlikely we’ll have any during the pilot.

Given it runs a query for seemingly every request it seems like
something we should remove until we have a use case for it.

As it’s very simple to install it’d be fast to adopt it if we do
have a fresh use case.