Basic view for admin homepage
This sadly replaces the amusingly basic admin of homepage of: "Hello
this is the admin interface".

I've tried to cobble together a few useful links and notices for an
admin user.

I set-up notice components to inform users of when chat is disabled and
when sign-ups are disabled. I did want to do them in a more aggressive
colour scheme than blue but didn't find a good component option that
would fit semantically.