Refactor unsubscribe system tests
This replaces the distinct unsubscribe system tests with scenarios in
the sign ups spec. The motivation for this is to make these tests more
thorough by working from the email rather than a route and to increase
consistency across our system tests.

I had to add in the unsubscribe link to the sign_in email so that the
test will work, however I know this will be shortly be removed as per
[1] and this may need rebasing depending on what is merged first.

There is also another conflict looming with [2] where we're splitting up
the sign-up and sign-in tests. So may need to rebase against that also.

[1]: https://github.com/alphagov/govuk-chat/pull/655
[2]: https://github.com/alphagov/govuk-chat/pull/652