Replace deprecated include: with includes:
This is to fix this warning (can't reproduce only saw it a couple of times)
```
warning: 299 OpenSearch-2.13.0-unknown "Deprecated field [include] used, expected [includes] instead"
```

The only information about this is based on Elasticsearch.

It makes no difference to the functionality of the app if you change it,
all the specs still pass, so both are supported, could be OpenSearch have decided
to hold off deprecating it or support both.

We should probably look at this again as the following blog post seems to
indicate we could get a performance boost by including the id in the document
when indexing a doing something like this:

```
  query: {
    term: { base_path: }
  },
  _source: false,
  stored_fields: "_none_",
  "docvalue_fields": ["id", "digest]
}
```

https://luis-sena.medium.com/stop-using-the-id-field-in-elasticsearch-6fb650d1fbae