Fix deprecation warning by changing sort order
The correct fix for this is to also duplicate the `_id` inside the document but
this would require re-indexing everything.

```
299 OpenSearch-2.13.0-unknown "Loading the fielddata on the _id field is
deprecated and will be removed in future versions. If you require sorting or
aggregating on this field you should also include the id in the body of your
documents, and map this field as a keyword field that has [doc_values] enabled"
```

As the sort order is irrelevant to the application, because we only use it to
create a hash, just changing it should fix things, again all specs are passing.

We should probably look at this again as the following blog post seems to
indicate we could get a performance boost by doing something like this:
```
GET test/_search
{
  "query": {
    "match_all": {}
  },
  "_source": false,
  "stored_fields": "_none_",
  "docvalue_fields": ["my_id_field"]
}
```
https://luis-sena.medium.com/stop-using-the-id-field-in-elasticsearch-6fb650d1fbae