Refactor function to take arguments rather than array
This refactors the `attachInputDescriptions` function to use the
arguments passed into it rather than expecting an array of values.

It's also a little confusing to require an empty array to be passed in
to "reset" the attribute.

Instead, we'll have two functions: one that takes any number of
arguments (DOM IDs) and sets those as teh input `aria-describedby`
attribute, along with the default `hintId` which we always want to set.

Then we have a resetter function which just sets the hintId.