Remove labels from metric route registration
Currently, the source label is being defined for all counters.

We don't need to actually define labels we pass in as evidenced by Pete's
work on adding metrics for answers. We send the status, question routing
label and guardrail output as part of this.

Interestingly, despite having the source label defined for the answer
counter metric it still works which shows that:

1. it's not a required label (we already knew this from my testing before)
2 you can pass different labels dynamically

We have 2 options:

1. always defined the label we want to pass so that devs can see which
labels are being passed
2. dynmically set them when we call the Metrics.increment_counter method

My vote is the latter as there's really nothing complex going. A dev will
easily be able to see what labels are being passed by looking at the
methods in the model that call the Metrics.increment_counter.

And it's also a bit misleading as it suggests you have to define the labels
on route registration when you don't.

I've tested removing the label works fine on integration.