Add question routing label enum
We have this enum defined at the database level, but not at the
application level.

This lets Rails handle the logic when a label is assigned to an answer
that doesn't appear in the database enum, as it'll not allow the answer
to be saved.

It also adds a spec to make sure that the enum here contains all of the
classifications listed in the LLM prompt config.