Question limit system message
This adds a system message to the chat history when the user is
approaching / has reached the question limit. So we show a message
when the user has 20 messages left, and one when they've reached the
limit.

For a JS request, the message gets appended to the `answer_html` string.
So the question is answered as normal, but we append some additional
HTML to this response which renders another message informing the user
about the limit.

Because it comes through via the fetch request, the message doesn't
appear when the page is refreshed, which is intentional.

We need to show this to non-JS users too, so we render the message on
the conversation show page if the remaining questions is 20 or 0.
Otherwise these users will never see the message.

This has the effect that if a JS-enabled user has exactly 20 questions
remaining, they'll see the system message in the message list when they
come to the chat. This trade-off feels acceptable though, in order to be
able to show the message to non-JS users.