Add conversation_total counter metric
This follows the prior art established for other counter metrics and
adds a counter for 'conversations_total'.

This passes one label 'first_conversation' which captures whether it's
the users first conversation or a subsequent one.

To avoid any issues on the Heroku app, i've ensured the
Conversation#send_conversations_total_to_prometheus returns if a
conversation doesn't have a user.