Add prometheus gauges for percentage of requests and tokens used
This commit adds guages for the percentage of rate limit metrics
used. This should make it easier to do follow up alerting on rate
limits. Previously gauges were added for the raw numbers of remaining
requests and tokens, but these are hard to act upon alone, as it's
less evident whether any particular value should be considered
problematic or not.