Raise in non-prod environments on undefined metrics
This PR is to resolve an issue we discovered in
https://github.com/alphagov/govuk-chat/pull/692 where nothing would
break if we didn't define a metric and we'd only see an issue once in
prod.

To resolve this I've added a conditional to raise an error rather than
report to Sentry in environments other than production. This is to have
the test suite provide visibility of any issues rather than waiting for
production.