Add custom headers for error/downtime endpoints
We've run into an issue with our Fastly configuration when we respond
with errors or downtime (uses a 503 or 410).

Fastly's 'vcl_error' subroutine catches the error and responds with a generic
error page https://github.com/alphagov/govuk-fastly/blob/816dc4bba5b10a5f5a12a361257c4f3cd7ce8ce8/www/www.vcl.tftpl#L686-L703
across GOV.UK.

We want to be able to serve users with our own error pages. We also want
to ensure that we don't serve the mirror (fallback) here
https://github.com/alphagov/govuk-fastly/blob/816dc4bba5b10a5f5a12a361257c4f3cd7ce8ce8/www/www.vcl.tftpl#L420

To get round this we can add 2 headers to our error respones:

"Govuk-Rendered-Error" - This will be used to identify when we want to
serve our own error page and allow us to call

```
 return (deliver);
```

within 'vcl_error' to move to the deliver hook and serve our error page.

The other header we need is the "No-Fallback" header. This is required
to ensure that we don't serve the mirror (fallback) page when 5xxs are
thrown.

This avoids restarts in this conditional https://github.com/alphagov/govuk-fastly/blob/816dc4bba5b10a5f5a12a361257c4f3cd7ce8ce8/www/www.vcl.tftpl#L686-L703

```
  if ((beresp.status >= 500 && beresp.status <= 599) && req.restarts < 3 && (req.request == "GET" || req.request == "HEAD") && !beresp.http.No-Fallback) {
    set beresp.saintmode = 5s;
    return (restart);
  }
```

'!beresp.http.No-Fallback' will return false and skip configuring saintmode
and restarting the process. Due to the presence of the header we know the
application is healthy and it's an intended response on our end.