Decrease homepage and static pages caching to 1 minute
The card mentions updating the homepage and static pages caching to 1 minute.
The reason for the homepage is if we toggle off the service 5 minutes
is too long before we serve the downtime page.

Since we continue to show the static pages when public access if off,
5 minutes of caching feels fine. I'll double check with Kev to see if
he agrees. But for now i'll update it to 1 minute.