Allow setting database pool size via env var
I realised that the number of connections in the database pool is lower
than our Sidekiq concurrency. Which means we're not following Sidekiq
config recommendations [1] and risking errors like:

```
ActiveRecord::ConnectionTimeoutError: could not obtain a connection from the pool within 5.000 seconds
```

To resolve this I'm adding an env var that we can use to configure the
size of the connection pool.

An alternative to setting an env var would be to have a pool that
matched our sidekiq concurrency (25) but doing so would mean using about
15 unnecessary connections per web worker, which would use a sizeable
chunk of the available connections.

[1]: https://github.com/sidekiq/sidekiq/wiki/Advanced-Options#concurrency