Reduce the Sidekiq default concurrency to 10
In the previous commit we are setting an env var to set the DB
connection pool size to allow us to match this to the sidekiq
concurrency in prod.

As we're going to set an env var in the helm charts for that we may as
well also use the same place to set the sidekiq concurrency.

Having this set to 25 by default, didn't prove ideal anyway with Heroku
needing to be set to a lower number. And there is of course unlikely any
need, outside of production, for us to be handling masses of questions.