Create table to store an audit of deleted users
This table is created with the intention that is populated whenever an
EarlyAccessUser is deleted. The intention is that this will provide us a
mechanism to look up data that would otherwise be lost.

Creating this feels a bit sub-optional as it's a bit of a halfway house
between a stats system and a soft delete function on early access user.
It's done this way as we're very near launch and this seems the quickest
approach - wouldn't be suprised if we consider removing it in future.