Create DeletedWaitingListUser table
As per the similar change two commits ago, this creates a table to store
a brief audit of when waiting list users are deleted so that we are able
to calculate the overall amount of users for analysis.

As per the deleted_early_access_users table, I'm not sure this type of
table is the best way to do this as it's halfway between a table of
stats and a soft delete but will suffice for now and a candidate for
iteration later.