Create a DeletedWaitingListUser when we delete a WaitingListUser
This is a near-identical process to the DeletedEarlyAccessUser system.
When we delete a WaitingListUser we create a DeletedWaitingListUser
record with a deletion_type field.

The motivation for this is to store a record of deletions so that we can
use the value in the DB for a full understanding of numbers on the
pilot.