Add user stats to admin homepage
This exposes counts of the amounts of users currently in the database
and number of users deleted.

This is to provide an overall insight into how many users have been
involved in the pilot.

As this relies on count queries there is a risk that this could become
slow if any of the tables involved becomes particularly large. I imagine
the queries that utilise enums could be sped up with indexes when there
is a lot of data.