Fix recording of embeddings rate limits
The json structure returned by OpenAI for embeddings calls is
different from that for chat completion, as a result embedding
rate limits weren't being picked up by the RateLimitMiddleware
and thus were not logged to Prometheus.

For simplicity, rather than dig into different json objects to
distinguish chat completions and embeddings, this commit changes
to using the endpoint url to distinguish whether an embadding or
chat model is being called.