Bump Ruby to 3.3.5
This is the version GOV.UK is now using in production so we should keep
in sync with it.

This update does produce a warning:

```
/root/.rbenv/versions/3.3.5/lib/ruby/3.3.0/bundled_gems.rb:75: warning: /root/.rbenv/versions/3.3.5/lib/ruby/3.3.0/ostruct.rb was loaded from the standard library, but will no longer be part of the default gems starting from Ruby 3.5.0.
You can add ostruct to your Gemfile or gemspec to silence this warning.
Also please contact the author of govuk_publishing_components-43.3.0 to request adding ostruct into its gemspec.
```

I've opened a PR to resolve: https://github.com/alphagov/govuk_publishing_components/pull/4251