Update questions filter to take conversation_id instead of conversation
Previously, when we had a conversation show page, we were grabbing the
id of the conversation from the path and passing that into the filter.

Going forward we're treating conversation like any other filter. The
conversation_id will be passed as a query string.

This updates the filter to take an optional conversation_id paramater.
If one is passed then it uses this to find the conversation.

It also updates the 'pagination_query_params' method to retain a
conversation_id when present in the pagination and filter links.