Configure routes.rb to only respond to supported formats
This changes the behaviour of the routes in the following ways:

- we no longer support routes such as /chat(.format) and only support
  paths without formats, for example chat.html will return a 404
- we default to a format of HTML
- routes that support multiple formats (html & json) require an accept
  header to request the format (as we already do)
- routes that support multiple formats will respond with 404s when
  requesting an unsupported format rather than a 406.

The motivation for this change started with a 500 logged in Sentry [1]
for a request of /chat.zip. I looked into this further and found things
we quite inconsistent across the app. Some routes would return an
unstyled 406 when receiving a request (those without explicit render
blocks and respond_to) and others returning 500. I felt this was better
as it reduces the number of routes that can return aspects of content
and stops weird edge cases (e.g. if you requested /chat.js you can get a
view rendered without a layout).

I had to fix a number of tests for this:

- it fixed routes that were requested with format as route argument
  rather than a request param
- resolved places where we testing JSON routes of routes that don't
  support JSON

I've added some basic and contrived tests around this that define
aspects of the behaviour. These are quite coupled to implementation. I
didn't want to set up tests that run on every route due to the
performance penalties that would have.

[1]: https://govuk.sentry.io/issues/5920989493/?environment=integration&environment=production&environment=staging&project=4507072589070336&query=is%3Aunresolved%20issue.priority%3A%5Bhigh%2C%20medium%5D&referrer=issue-stream&statsPeriod=90d&stream_index=1