Store LLM response for question routing field
This fixes a bug that I spotted where an error on a JSON schema was
showing as the llm_response not the question_routing_llm_response field.

While looking at this I noticed we were also not converting the input to
JSON and putting a hash into the Answer model, inconsistently with the
other models. I figured this could be because we're worried the JSON may
not be valid and thus we don't have a hash as a value - however the code
prior to JSON parse also relies on this. Presumably if we get a response
from OpenAI that isn't valid JSON we'd get a different error as it'd
raise an exception at the Faraday request.