Assign LLM response when running guardrailing
We're moving from storing the LLM response in separate fields to storing
it in one JSONB field. This updates the guardrail generation step of the
pipeline to call `answer.assign_llm_response` instead of assigning it as
an attribute.

The `FewShot` class would only return the full message content as
`llm_response`, so now we're returning the whole `choices.[0]` hash now
so that we can store the entire response against the answer.