Extract OpenAI tool calling assertion into method
In these tests we want to assert that the LLM response has been stored
against the answer.

It's overkill to match the entirety of the response as it contains a lot
of data, but we want to verify that we're matching a response that
contains a tool call.

To reduce duplication, this matching has been extracted into a method
so we can assert the name of the tool call in the spec without having to
repeat ourselves so much.