Metrics dashboard for admin area
This adds a dashboard that uses Chartkick to visualise a number of
insights into GOV.UK Chat usage over the previous 7 days. It is not
available without JS (as is similar with tools such as Grafana or Looker
studio). As we have very limited access to the admin area I have not
made any counter measures for lack of JS support.

I've taken the approach of loading the data for each dashboard over JSON
rather than rendering it in a single page view. Doing this via JSON
offers a few advantages:

- page rendering is not dependent on slowest metric
- metrics are auto refresh without page reload
- easier to spot which metric is the source of any performance issues
- easier to test the data

I left the individual JSON routes as supporting html and json so that
when you opened them in a browser you'd be served the JSON (otherwise
you'll get a 404 as the browser would request html)

As the data is generated using group by queries there is a risk of
incomplete data sets for when there is no activity (i.e. if you query
across 7 days but only 3 days have activity you'd only get 3 days of
data back). So I had to write a function that'd ensure the expected days
were there.

Chartkick makes use of inline styles so I had to add an exception to the
content security policy to allow this.

I put together a relatively simple system for test that checks merely
for the presence of a canvas element (this is produced when the JS has
been able to load the JSON and generate a chart). I could not establish
a means to assert what data is shown in the chart.