Set 'Try GOV.UK Chat' as h1 on both start page variations
This was raised by content and design as the h1 should be the same across both start page variations (the Figma designs which indicated dynamic text should be used were out of date). Since we're only setting the h1 content in one place here, I opted for removing the title_text variable and just hardcoded the text - happy to discuss if we want to retain the variable.