Include output guardrail failures on admin metrics
This metric is used to show the occurrence of each individual output
guardrail that is triggered. This is a harder metric to extract than the
other ones because more than output guardrail can trigger at a time and
we store them in an array table.

We can't use the GROUP BY SQL statement to get the grouping for each
individual one (well at least I can't others may have a technique). So
I've wrote a class method that can be applied to a grouped scope to
separate out the output guardrail failures into individual occurrences.