Remove prometheus count metrics
We've had a bit of a miserable time trying to turn these metrics into
any valuable insights and currently don't have confidence in them. As
we've not got a current usage for them it seems simplest to remove them.

We can always consider adding data back once we have a wider idea of
usage - probably in a more gradual way than last time.