Add list of forbidden terms to config
Adds a list of forbidden words to be used in a last resort check to
abort any LLM answer message that contains inapprioriate, profane
and/or offensive terms.

The list contains words or short phrases
which would always be inappropriate in the context of GOV.UK Chat.
Words which are only inappropriate in specific contexts are not
included.