Fix invalid JSON schema for structured answer
We had additionalProperties and required outside of the parameters
definition which was making this schema invalid. This change wasn't
noticeable until we had strict mode on as OpenAI was just doing it's
best with it. However once strict mode was enabled we'd get errors back
from OpenAI such as:

```
Invalid schema for function 'generate_answer_using_retrieved_contexts': In context=(), 'additionalProperties' is required to be supplied and to be false.
```