Override govspeak ordered list styles for conversation messages
This applies an override to the styling of ordered lists at the request
of the design part of GOV.UK Chat. The reason to do this override is
because Safari has a styling issue with list markers [1] which has been
flagged as an issue with the GOV.UK Design System team [2].

This has an effect in Safari that list markers are centre aligned along
the middle of the paragraph left margin, whereas in other browsers the
list markers are left aligned next to that margin. This has the visual
effect that the markers are breaking out of their margin - particularly
evident once a list marker is in double figures.

This fix carries a number of risks with it as it won't work for all
ordered list styles (however it is unlikely, as the source data is
markdown, that we'll have complex ones) and could easily be broken by
changes to underlying styles in ways that our automated testing won't
detect.

I've also added an example to our component guide so it is easy to
witness this issue.

[1]: https://github.com/alphagov/govuk-chat/pull/761
[2]: https://github.com/alphagov/govuk-frontend/issues/5398

Co-authored by: Kevin Dew <kevindew@me.com>