Set max completion tokens to 160 for question routing
We've done some research on the average and max token completions it takes
when the question router categorses a question as
vague_acronym_grammar, advice_opinions_predictions or multi_questions.

We've found that it's ~130. We want to limit the output to a number above
this to ensure we're not cutting off legitmate completions, but also
be able to control the output of prompt injection attacks.

If the completion token limit is reached, it returns incomplete JSON in
the answer property.

We can rescue this and update the answers status to an abort and
set the question routing label to the correct category.