Question routing canned responses
This adds canned responses for the question routing labels which we want
to return early from.

When a user asks a question which gets classified as any of these
labels, we want to return a random canned response from the available
list. So a method has been added to the `Answer::CannedResponse` module
to return a random element.