Update question router to return LLM answer for selected classifications
Currently the question router will make a call to the LLM which
includes a whole array of tool calls. Each tool call represents a
classification (label) which the LLM can choose from to apply to the
question.

All of the tool calls will instruct the LLM to return two properties in
its JSON response:

1. an answer
2. a confidence value

Getting it to come up with its own answer - which we will in turn present
to the user - leaves it open to prompt injection.

To minimise this issue, we want to return a canned response to the user
if the LLM classifies their question in a certain way. For other
classifications, we want to carry on as normal (and we'll eventually run
guardrails against the answer the LLM produced).

So this commit updates the logic for the question router to do just
this. The `properties` attribute has been removed in most of the tool
calling functions, meaning the LLM will return JSON with only the
confidence value, and no `answer` attribute.

If we don't see an `answer` attribute in the response then we abort the
pipeline and select a random canned response for that particular
classification. (The canned responses are defined in the question
routing labels YAML config).

If we do see an `answer` attribute, or if the classification is
`genuine_rag`, then we continue along with the pipeline.