Fix guardrail evaluation rake task
This rake task runs the guardrails against an evaluation CSV file.

It should return a string of the actual guardrail response in the
`actual` attribute of the JSON - e.g. `'True | "1"'`

A change was made to the `FewShot` class recently to the `llm_response`
attribute. Previously it would return just the string mentioned above.
The change made it return the entire LLM response so that it could be
persisted in the database.

This messed up the evaluation rake script because it generated the JSON
using this `llm_response` attribute, so it went from rendering the above
string to rendering the whole LLM response.

This fixes that issue. We'll keep the `llm_response` attribute of the
`FewShot::Result` class to be the entire LLM response, and we'll add a
new attribute, `llm_guardrail_result` which is just the above string.

We could have left this `Result` class alone and instead dug into the
`llm_response` hash in the rake file to get the content we want, but I
think it's probably nicer to expose the string as an attribute so you
don't need to know the internals of an OpenAI API response to get the
data you're interested in.