Allow FewShot guardrail class to use different prompts
We'll be adding some logic to run guardrails against the question
routing answer soon.

The actual guardrailing process will be exactly the same as the current
output guardrails, with the exception of the prompt being different.

This commit modifies the `FewShot` class so that the prompt name is
injected in the initializer rather than being hardcoded in the class.