Add question routing guardrails pipeline class
This class acts pretty similarly to the output guardrails class but has
a few logical differences.

Firstly, the prompt is different, so we need to inject the prompt name
into the `FewShot` class when calling it.

If the question routing label produced in the previous step of the
pipeline (`QuestionRouter`) is `genuine_rag` then we don't need to run
the guardrailing at all. In this instance, the LLM hasn't produced an
answer to return to the user, so there's no need to check it. In this
situation, we just return immediately.

If not `genuine_rag` then we run the guardrails using the new prompt.

The key difference between this and the output guardrailing class is
that here we _always_ want to abort the pipeline. We know the question
the user asked isn't `genuine_rag` so we never want to continue the
pipeline after this step. For that reason, we always sort-abort the
pipeline here (i.e. don't use the `abort_pipeline!` bang method.

If the guardrails have triggered - i.e the answer produced by the LLM in
the question routing stage is unsuitable for whatever reason - then we
assign a canned response to the answer in the context. This is the
answer the user will see.

If the guardrails aren't triggered, then we can safely show the user the
answer produced by the LLM in question routing. This is already assigned
to the `context.answer.message` attribute so we don't need to do
anything.

We then assign the relevant statuses and guardrail failures (if any).