Jailbreak guardrails
This applies a guardrail written by Alessia to determine whether a user
question may be a form of jailbreak. This applies the code that
surrounds this.

The code for this is currently in a `InputGuardrails` namespace where we
plan to move this into a `Guardrails` directory once `OutputGuardrails`
is renamed to that.

I've included all aspects of the prompt instruction in the YAML file
rather than aspects of it in the YAML and in the code and tests. This
was done so that when we make the code for this public and the prompt
private that we don't bleed into the code what is needed to pass this
guardrail or not.

Co-Authored by: Alessia Tosi <alessia.tosi@digital.cabinet-office.gov.uk>
Co-Authored by: Kevin Dew <kevindew@me.com>