Dynamic max_token parameter for guardrails
The guardrailing class reads in its config from a YAML file. This file
defines the rules by which the LLM determines whether an input should
pass or fail.

Currently we specify the `max_tokens` as a constant. We can do this
because we know that there's a limit on the string that the LLM will
return to us. It'll be either:

1. "False"
2. "True | [a numerical list of the failures]"

So the maximum length of string can be calculated as "True" followed by
all of the mappings.

But given the mappings are dynamic, depending on the config, this
constant shouldn't be hardcoded.

Instead we can dynamically calculate it by building up the worst-case
string, tokenizing it and adding a small buffer.