Rename output_guardrail_status to answer_guardrails_status
This makes it clearer that this value relates to guardrailing the answer
produced by the LLM.

We'll be soon introducing a question routing guardrail step, so this
helps reduce confusion as "output_guardrails" will be too generic a
name.

It also makes naming consistent by using the plural form - i.e
"guardrails" rather than "guardrail". We'll be using the plural form
for everything from now on.