Add answer_guardrails enum values
This adds two enum values which are a direct replacement for the
existing values representing error/abort for guardrails.

We're renaming "output" guardrails to "answer" guardrails, so this
adds the enum values to support that.

We can't yet remove the old enum values as there might still be records
referencing them.