Remove jailbreak attemps from message history
We have the following potential vunerability:

- User inputs a jailbreak
- System detects it and rejects
- User sends a message of “try again”
- System doesn’t detect “try again” as a jailbreak
- Question rephraser combines “try again” with the previous jailbreak question.

To get round this we need to remove the jailbreak attempts from the message
history passed to the question rephraser.

We also need to update how we identify if it's the users first question.
Previously we could just check if conversatons question count was 1
(just the question built by the context on init).

Now we need to get the last 5 questions that have jailbreaks filtered out
and check if any are present. If not we can return early.