Fix context nesting in few shot spec
All of the specs in this file were nested under the "when the request is
successful" context.

This move the "with a question routing guardrail LLM prompt" into the
successful request context, as it's describing a successful request.

Then it moves the two contexts asserting the OpenAI API errors outside
of the success context.