Remove re-raises of OpenAI exceptions
This removes the places where we re-raise OpenAI exceptions with
modified messages as found in guardrails and question rephrasing. These
have been removed because I don't believe they are adding extra value.
This has been done to be consistent with https://github.com/alphagov/govuk-chat/pull/767
that skipped them.

Initially we added these on with an expectation that they may make it
easier to debug a problem, but that hasn't proven to be case and the
errors are very rare.